feat: Add MovieJSONResult class
This class generates the parent object from the returned JSON object
when retrieving a page of movies from the TMDB API.  The main feature
is the results List, which contains a list of movies from the given
page.  This object is instantiated through deserialization of the
JSON response via Gson.