refactor: Set OkHttpClient maxRequests to 1
Set up the OkHttpClient to be a static member variable of the
RemoteMovieDataSource class, which allows all calls to
getJsonFromRequest() to access the same client object.

From here, create a new method, initializeClient(), which
instantiates the object, sets the dispatcher's maxRequests
to 1, and sets an idleCallback which sets the client to
null to prevent any potential memory leaks.

Setting maxRequests to 1, when by default it is set to 64,
ensures that only 1 call is being processed at a time, while
still allowing additional calls to be queued.  This ensures
that the calls are being executed in order of page number,
to better handle multiple calls to the API.