feat: Add new refreshMovies() method
This method works with the new cached movieList from the
MovieRepository.  In the event that the presenter's start
method is called, but the firstLoad boolean is false,
this will check the repository to see if there are cached
movies.  If there is not, then load the first page of movies
as normal.  This saves time and resources and restores the state
of the grid view to its original state.