feat: Add automatic calls to loadMovies()
Using a custom MovieScrollListener for the RecyclerView,
detect when the user is close to reaching the end of the
currently loaded grid of movies.  When this occurs, call
loadMovies while blocking additional calls until the load
is complete with the readyToLoad boolean flag.  With this
in place, continuous scrolling can be better achieved.