fix: Better define condition for movieView.displayNoMovies()
Add method to MovieRepository that returns cached movie list size.
If this is 0, there are no movies to be displayed, at which the
presenter can then order the view to display the error layout.
The previous usage of checking the current page number was broken,
since this is automatically incremented to > 1 after the initial
loadMovies() call.