refactor: Implement better handling of API page number
Move responsibility of managing TMDB API current page number from
MoviePresenter to RemoteMovieDataSource.  In the previous implementation,
it was the presenter's job to know what page was being asked for, and to
increment/decrement/reset as necessary.  When a request for getMoreMovies was issued,
the page number variable was immediately incremented, regardless of callback response.
This was hard to manage with asynchronously queued requests,
especially in instances of network failures, where the expected behavior is that
if page x failed to load, page x would be attempted again in the next request, not
page (x + 1).

To better handle this, the RemoteMovieDataSource now increments its own
page number variable when the OkHttp request's onResponse callback is
called.  If onFailure is called, we don't change the page number, allowing
the next getMoreMovies request that comes through to attempt that page
number again.

Finally, in the instance of user selecting a new sort option, the expected
behavior is that the movie cache will clear out and the next API request
will be for page 1.  To handle this, implement a boolean parameter to the
MovieDataSource interface's getMoreMovies function and add a newSortSelected
member variable to the MoviePresenter.  This variable will be set to true
when the user calls the presenter's setSortParameter method, and immediately
back to false after the consequent loadMovies call.  The RemoteMovieDataSource
then checks to see if this parameter is true, in which case it resets the
page number to 1.