fix: Dismiss displayed error snackbar on displayNewMovies
Make the connection error snackbar a member variable that is
instantiated in onCreate.  When displayLoadMoviesError is called
this variable can be referenced from here.  The problem was that,
since the duration is set to indefinite, the message will be dismissed
if the network response is successful after the user clicks "RETRY",
but not if the user swipes up and down to trigger the scroll listener
trigger.  The expected behavior is that this message will be dismissed
after a successful response, regardless of how it was initially called for.
With the new member variable set, the displayNewMovies now checks
to see if the snackbar is displayed, and if so, dismisses it.