fix: Better handle retrieval and displaying of cached movies
When navigating away and returning to MovieActivity, onResume is
called, which then calls the presenter's refreshMovies method.
This was causing the cached list to be retrieved and added onto
the adapter's existing list, causing duplicate movie entries.
Fix this by adding displayCachedMovies method, which resets
the adapter list and sets it to the cached list, to ensure
that the adapter list always reflects the cached list,
and to ensure that duplicates are no longer being displayed.