refactor: Set home back button to finish()
Override onOptionsItemSelected to set user clicks of the back arrow
home button to call finish() instead of the superclass implementation.
This destroys the DetailActivity and returns the user to the
MovieActivity grid position that they left off at.  Without this,
the MovieActivity starts back at the top, which is not desirable,
especially if the grid list gets long.