refactor: Set Movie to Implement Parcelable
Implement Parcelable for the Movie entity to facilitate the transfer of
Movie data between the MovieActivity and DetailActivity.  In the new setup,
both the parcelled Movie object and its movie ID are passed to the DetailActivity
with the Intent, which is then passed to the DetailPresenter.  If the
Movie object received by the presenter is not null, it will use this to
populate the DetailActivity UI.  If it's null, it will then use the movie ID
to retrieve the Movie details from the MovieRepository.