refactor: Add functionality to retrieve movie by ID from the TMDB API
In order to improve the robustness of the DetailActivity.  In the event that
the Movie parcelable object is null, and the MovieRepository does not have
the movie cached when searching by movie ID, the MovieRepository will fetch
the movie from the RemoteMovieDataSource by ID.

The required creating a new API request flow and a new LoadMovieCallback, as well
a new List<String> genreList member variable in the Movie entity.  The JSON
result for a TMDB get movie by ID request returns a list of Genre objects rather
than merely Genre IDs, like in the get movies by popularity/highest-rated requests.
Add handling to support this to properly populate the field when necessary and
to retrieve and present the data in the Presenter.