refactor: Better handle calls to repo getMovieById
Add a new private loadMovies(Movie, GenreMap) function.  This
function is called within the overridden loadMovies function
either when the initial check to see if the movie is null
returns false, or if true, when the repository has
retrieved the movie data via getMovieById's callback.
With this new function in place, we can remove the ugly
boolean flag and while loop, which was intended to halt the
progress of the loadMovies method if an asynchronous call to
getMovieById was necessary.