feat: Add Room DB for local Favorite Movies
Add functionality to support Movie objects to be stored in local Room DB
to track user's marked favorite movies.  This table will only store the
Movie's ID and Poster Path values.  That way, the poster image can be
cached and easily displayed in the MovieActivity grid when viewing
movies sorted by Favorites.  When the user clicks a poster, the
presenter will then fetch the movie object from the repository's
remote data source in order to ensure that the variable vote average and
any other variable data is always up to date.
such as vote average and